fix: query 수정
