+++ src/state/mutations/time/usePutTimeVoteMutation.ts
@@ -24,8 +24,12 @@ export const usePutTimeVoteMutation = (
       putTimeVote({ roomId, dateTime }),
     onSuccess: () => {
       queryClient.invalidateQueries({
-        queryKey: TIME_KEY.GET_TIME_VOTED(roomId!),
+        queryKey: [TIME_KEY.GET_TIME_VOTED(roomId!)],
       });
+      queryClient.invalidateQueries({
+        queryKey: [TIME_KEY.GET_TIME_RESULT(roomId!)],
+      });
+
       navigate(PATH.TIME_RESULT(roomId));
     },
     ...options,
